Add convenience functions for array.
Allow to pass std::vector<T> for get(), as(), and is().
For example:
v.get<std::vector<int>>("key")
v.as<std::vector<int>>("key")
v.is<std::vector<int>>()

Note that we see only the first element of the array for testing
the element type of array. If an array contains multiple types of
element, exception will be raised when converting toml::Value to
the specified type.

Fixed #7.